app/requirements.txt
flask
flask-mysqldb
flask-jwt-extended
